Connect4Animation Updates
Added changes to Connect4Animation in attempts to make it a glasspane,
still yet to test. Also came to realisation that probably do not need to
get JFrame and BoardGame in constructor, so removed them, left them in
Superclass for now incase Ben needs them, but believe it best they be
removed and JFrame and BoardGame from GUI used instead.